vanno aggiunto mqtt e sitemata la parte delle richieste in interfaccia

+++ WaterManagerUI/Pages/Admin/Elimina/EliminaEsigenza.cshtml.cs
@@ -24,35 +24,41 @@ public class EliminaEsigenza : PageModel
 
     public async Task OnGetAsync()
     {
-        var client = _httpClientFactory.CreateClient();
-        client.DefaultRequestHeaders.Authorization =
-            new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
-        var response = await client.GetAsync("http://localhost:8080/api/v1/utils/esigenza/get/all");
-
-        if (response.IsSuccessStatusCode)
+        if (_signInManager.IsSignedIn(User) && User.FindFirstValue(ClaimTypes.Role).Equals("SYSTEMADMIN"))
         {
-            var content = await response.Content.ReadAsStringAsync();
-            this.esigenze = JsonConvert.DeserializeObject<HashSet<string>>(content);
+            var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization =
+                new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+            var response = await client.GetAsync("http://localhost:8080/api/v1/utils/esigenza/get/all");
+
+            if (response.IsSuccessStatusCode)
+            {
+                var content = await response.Content.ReadAsStringAsync();
+                this.esigenze = JsonConvert.DeserializeObject<HashSet<string>>(content);
+            }
         }
     }
 
     public async Task<IActionResult> OnPostAsync()
     {
-        var client = _httpClientFactory.CreateClient();
-
-        if (_signInManager.IsSignedIn(User))
+        if (_signInManager.IsSignedIn(User) && User.FindFirstValue(ClaimTypes.Role).Equals("SYSTEMADMIN"))
         {
-            try
-            {
-                client.DefaultRequestHeaders.Authorization =
-                    new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+            var client = _httpClientFactory.CreateClient();
 
-                await client.DeleteAsync($"http://localhost:8080/api/v1/admin/esigenza/delete/{esigenza}");
-            }
-            catch (Exception ex)
+            if (_signInManager.IsSignedIn(User))
             {
-                Console.WriteLine(ex.StackTrace);
-                return RedirectToPage("/Admin/Coltivazione/Esigenza");
+                try
+                {
+                    client.DefaultRequestHeaders.Authorization =
+                        new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+
+                    await client.DeleteAsync($"http://localhost:8080/api/v1/admin/esigenza/delete/{esigenza}");
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.StackTrace);
+                    return RedirectToPage("/Admin/Coltivazione/Esigenza");
+                }
             }
         }
