+++ WaterManagerUI/Pages/Admin/Elimina/EliminaRaccolto.cshtml.cs
@@ -12,7 +12,7 @@ public class EliminaRaccolto : PageModel
     [BindProperty] public String raccolto { get; set; }
     private readonly IHttpClientFactory _httpClientFactory;
     private readonly SignInManager<IdentityUser> _signInManager;
-    public HashSet<string> irrigazioni { get; set; }
+    public HashSet<string> raccolti { get; set; }
 
     public EliminaRaccolto(IHttpClientFactory httpClientFactory, IHttpContextAccessor httpContextAccessor,
         SignInManager<IdentityUser> signInManager)
@@ -24,35 +24,41 @@ public class EliminaRaccolto : PageModel
 
     public async Task OnGetAsync()
     {
-        var client = _httpClientFactory.CreateClient();
-        client.DefaultRequestHeaders.Authorization =
-            new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
-        var response = await client.GetAsync("http://localhost:8080/api/v1/utils/Raccolto/get/all");
-
-        if (response.IsSuccessStatusCode)
+        if (_signInManager.IsSignedIn(User) && User.FindFirstValue(ClaimTypes.Role).Equals("SYSTEMADMIN"))
         {
-            var content = await response.Content.ReadAsStringAsync();
-            this.irrigazioni = JsonConvert.DeserializeObject<HashSet<string>>(content);
+            var client = _httpClientFactory.CreateClient();
+            client.DefaultRequestHeaders.Authorization =
+                new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+            var response = await client.GetAsync("http://localhost:8080/api/v1/utils/raccolto/get/all");
+
+            if (response.IsSuccessStatusCode)
+            {
+                var content = await response.Content.ReadAsStringAsync();
+                this.raccolti = JsonConvert.DeserializeObject<HashSet<string>>(content);
+            }
         }
     }
 
     public async Task<IActionResult> OnPostAsync()
     {
-        var client = _httpClientFactory.CreateClient();
-
-        if (_signInManager.IsSignedIn(User))
+        if (_signInManager.IsSignedIn(User) && User.FindFirstValue(ClaimTypes.Role).Equals("SYSTEMADMIN"))
         {
-            try
-            {
-                client.DefaultRequestHeaders.Authorization =
-                    new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+            var client = _httpClientFactory.CreateClient();
 
-                await client.DeleteAsync($"http://localhost:8080/api/v1/admin/raccolto/delete/{raccolto}");
-            }
-            catch (Exception ex)
+            if (_signInManager.IsSignedIn(User))
             {
-                Console.WriteLine(ex.StackTrace);
-                return RedirectToPage("/Admin/Coltivazione/Raccolto");
+                try
+                {
+                    client.DefaultRequestHeaders.Authorization =
+                        new AuthenticationHeaderValue("Bearer", User.FindFirstValue(ClaimTypes.Authentication));
+
+                    await client.DeleteAsync($"http://localhost:8080/api/v1/admin/raccolto/delete/{raccolto}");
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.StackTrace);
+                    return RedirectToPage("/Admin/Coltivazione/Raccolto");
+                }
             }
         }
